% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{centering_function}
\alias{centering_function}
\title{centers the data}
\usage{
centering_function(data, method_of_centering, grouping_vector,
  to_standardize = F)
}
\description{
centers the data
}

