% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{cluster_freq_function}
\alias{cluster_freq_function}
\title{calculates cluster centroids}
\usage{
cluster_freq_function(data, n_clusters, kfit, variable_names)
}
\description{
calculates cluster centroids
}

