% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{cluster_plot_function}
\alias{cluster_plot_function}
\title{creates plot of cluster centroids}
\usage{
cluster_plot_function(cluster_freqs, font_size, fill_order)
}
\description{
creates plot of cluster centroids
}

