% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_functions.R
\name{compare_cluster_statistics}
\alias{compare_cluster_statistics}
\title{Function to compare the proportion of variance explained for cluster solutions with varying number of clusters}
\usage{
compare_cluster_statistics(prepared_data, args, lower_num, upper_num)
}
\arguments{
\item{cluster_assignments}{cluster assignments from calculate_stats() function, in particular the fifth list item from its output}
}
\description{
Function to compare the proportion of variance explained for cluster solutions with varying number of clusters
}
\details{
Function to compare the proportion of variance explained for cluster solutions with varying number of clusters
}

