% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{comparision_of_statistics_plot}
\alias{comparision_of_statistics_plot}
\title{compare cluster fit index statistics}
\usage{
comparision_of_statistics_plot(data, lower_num, upper_num)
}
\description{
compare cluster fit index statistics
}

