% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_compare_anova}
\alias{create_compare_anova}
\title{compares between levels of factor using ANOVA}
\usage{
create_compare_anova(processed_data, variable_to_find_proportion, cluster_names,
  factor_to_explore)
}
\description{
compares between levels of factor using ANOVA
}

