% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_plot_to_explore_factors}
\alias{create_plot_to_explore_factors}
\title{create plot to explore factors}
\usage{
create_plot_to_explore_factors(processed_data, factor_to_explore, cluster_names)
}
\description{
create plot to explore factors
}

