% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_processed_data}
\alias{create_processed_data}
\title{creates process data for summary table and plot}
\usage{
create_processed_data(raw_data, factor_to_explore, variable_to_find_proportion)
}
\description{
creates process data for summary table and plot
}

