% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_raw_data}
\alias{create_raw_data}
\title{create raw data for ANOVA and subsequent processing}
\usage{
create_raw_data(dummy_coded_data, factor_to_explore,
  variable_to_find_proportion)
}
\description{
create raw data for ANOVA and subsequent processing
}

