% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_functions.R
\name{explore_factors}
\alias{explore_factors}
\title{Function to explore frequency of clusters across select factors}
\usage{
explore_factors(cluster_assignments, cases_to_keep, factor_data_frame,
  factor_to_explore, variable_to_find_proportion = NULL,
  cluster_names = NULL, variable_names = NULL, print_status = T)
}
\arguments{
\item{cluster_assignments}{cluster assignments from calculate_stats() function, in particular the fifth list item from its output}

\item{cases_to_keep}{cases to keep from calculates_stats() function, in particular the attribute "cases_to_keep" from its output}

\item{factor_data_frame}{data frame of select factors}

\item{factor_to_explore}{specific factor to explore}

\item{variable_to_find_proportion}{variable to normalize clusters as a unit of analysis}

\item{cluster_names}{optional names for clusters, useful for interpreting findings}

\item{variable_names}{optional names for variables that were clustered}

\item{print_status}{boolean indicating whether to print information about the output (TRUE) or to not print information about the output (FALSE)}
}
\description{
Function to explore frequency of clusters across select factors
}
\details{
To explore the frequency of clusters across factors
 and cleaning the corpus, deviationalizing and clustering.
}

