% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{find_nearest_centroid}
\alias{find_nearest_centroid}
\title{finding the nearest centroids}
\usage{
find_nearest_centroid(split_halves, calculated_stats)
}
\description{
finding the nearest centroids
}

