% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{hclust_to_kmeans_function}
\alias{hclust_to_kmeans_function}
\title{converts hclust output to start values for kmeans function}
\usage{
hclust_to_kmeans_function(data, out, n_clusters)
}
\description{
converts hclust output to start values for kmeans function
}

