% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{merge_assignments_and_factors}
\alias{merge_assignments_and_factors}
\title{merge assignments with selected factors}
\usage{
merge_assignments_and_factors(cluster_assignments, cases_to_keep,
  factor_data_frame)
}
\description{
merge assignments with selected factors
}

