% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_clusters.R
\name{plot_profiles}
\alias{plot_profiles}
\title{Return plot of profile centroids}
\usage{
plot_profiles(d, to_center = F, to_scale = F)
}
\arguments{
\item{d}{summary data.frame output from create_profiles_cluster()}

\item{to_center}{whether to center the data before plotting}

\item{to_scale}{whether to scale the data before plotting}
}
\value{
A ggplot2 object
}
\description{
Return plot of profile centroids
}
\details{
Returns ggplot2 plot of cluster centroids
}
