% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_r_squared}
\alias{plot_r_squared}
\title{Plot R^2 (r-squared) values for a range of number of profiles}
\usage{
plot_r_squared(df, ..., to_center = FALSE, to_scale = FALSE,
  distance_metric = "squared_euclidean", linkage = "complete",
  lower_bound = 2, upper_bound = 9, r_squared_table = FALSE)
}
\arguments{
\item{df}{with two or more columns with continuous variables}

\item{...}{unquoted variable names separated by commas}

\item{to_center}{(TRUE or FALSE) for whether to center the raw data with M = 0}

\item{to_scale}{Boolean (TRUE or FALSE) for whether to scale the raw data with SD = 1}

\item{distance_metric}{Distance metric to use for hierarchical clustering; "squared_euclidean" is default but more options are available (see ?hclust)}

\item{linkage}{Linkage method to use for hierarchical clustering; "complete" is default but more options are available (see ?dist)}

\item{lower_bound}{the smallest number of profiles in the range of number of profiles to explore; defaults to 2}

\item{upper_bound}{the largest number of profiles in the range of number of profiles to explore; defaults to 9}

\item{r_squared_table}{if TRUE, then a table, rather than a plot, is returned; defaults to FALSE}

\item{n_profiles}{The specified number of profiles to be found for the clustering solution}
}
\value{
A list containing a ggplot2 object and a tibble for the R^2 values
}
\description{
Plot R^2 (r-squared) values for a range of number of profiles
}
\details{
Returns ggplot2 plot of cluster centroids
}
\examples{
df <- mtcars
plot_r_squared(df, mpg, wt, hp, qsec, to_center = TRUE, to_scale = TRUE, lower_bound = 2, upper_bound = 4)
}
