% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_functions.R
\name{prepare_data}
\alias{prepare_data}
\title{Pre-processing function to prepare data for subsequent analysis}
\usage{
prepare_data(raw_data, method_of_centering = "raw", grouping_vector = NULL,
  to_standardize = F)
}
\arguments{
\item{raw_data}{data frame or matrix of any dimensions with numeric data}

\item{method_of_centering}{string indicating variable-wise centering, options include "grand" (for grand mean centering), "group" (for group mean centering; requires a grouping vector, described next) and "raw", which does not center the variables}

\item{grouping_vector}{a vector indicating how the cases are to be grouped for group mean centering}

\item{to_standardize}{boolean indicating whether to standardize (TRUE) or not (FALSE)}
}
\description{
Pre-processing function to prepare data for subsequent analysis
}

