% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.prcr}
\alias{summary.prcr}
\title{Concise summary of prcr cluster solution}
\usage{
\method{summary}{prcr}(x)
}
\arguments{
\item{x}{A `prcr` object}
}
\description{
Concise summary of prcr cluster solution
}
\details{
Prints a concise summary of prcr cluster solution
}
