% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_functions.R
\name{calculate_stats}
\alias{calculate_stats}
\title{Function to calculate statistics about cluster solution found via cluster_data()}
\usage{
calculate_stats(clustering_output, variable_names = NULL,
  cluster_names = NULL)
}
\arguments{
\item{clustering_output}{output from cluster_data() function}

\item{variable_names}{optional names for variables that were clustered}

\item{cluster_names}{optional names for clusters, useful for creating plot}
}
\description{
Function to calculate statistics about cluster solution found via cluster_data()
}

