% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_functions.R
\name{calculate_stats}
\alias{calculate_stats}
\title{Function to calculate statistics about cluster solution found via cluster_data()}
\usage{
calculate_stats(clustering_output, names_of_clusters = NULL)
}
\arguments{
\item{clustering_output}{output from cluster_data() function}

\item{names_of_variables}{optional names for clusters, useful for interpreting findings}
}
\description{
Function to calculate statistics about cluster solution found via cluster_data()
}

