% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_functions.R
\name{compare_cluster_statistics}
\alias{compare_cluster_statistics}
\title{Function to compare the proportion of variance explained for cluster solutions with varying number of clusters}
\usage{
compare_cluster_statistics(prepared_data, args, lower_num, upper_num)
}
\arguments{
\item{prepared_data}{output from the prepare_data() function}

\item{args}{list of arguments passed to the create_profiles() function, which is returned as an attribute of its output (i.e., attributes(output)$args_attr)}

\item{lower_num}{integer representing the lower bound of the range of the number of clusters in the cluster solution}

\item{upper_num}{integer representing the lower bound of the range of the number of clusters in the cluster solution}
}
\description{
Function to compare the proportion of variance explained for cluster solutions with varying number of clusters
}
\details{
Function to compare the proportion of variance explained for cluster solutions with varying number of clusters
}

