% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_functions.R
\name{create_profiles}
\alias{create_profiles}
\title{Create profiles function}
\usage{
create_profiles(prepared_data, n_clusters,
  distance_metric = "squared_euclidean", linkage = "complete")
}
\arguments{
\item{prepared_data}{output from the prepare_data() function}

\item{n_clusters}{the number of clusters; specified a priori}

\item{distance_metric}{metric for calculating the distance matrix used in hierarchical clustering, options include "euclidean", "squared_euclidean", and others (see ?dist() for more details)}

\item{linkage}{method for combining clusters in hierarchical clustering, options include "complete", "average", and others (see ?hclust() for details)}
}
\description{
Create profiles function
}

