% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_new.R
\name{create_profiles}
\alias{create_profiles}
\title{# Create profiles}
\usage{
create_profiles(df, ..., n_profiles, model_name = "VVV", center_data = TRUE,
  scale_data = TRUE, distance_metric = "euclidean", linkage = "complete",
  to_return = "tibble")
}
\arguments{
\item{df}{with two or more columns with continuous variables}

\item{...}{unquoted variable names separated by commas}

\item{distance_metric}{Distance metric to use for hierarchical clustering; "squared_euclidean" is default but more options are available (see ?hclust)}

\item{linkage}{Linkage method to use for hierarchical clustering; "complete" is default but more options are available (see ?dist)}

\item{to_center}{(TRUE or FALSE) for whether to center the raw data with M = 0}

\item{to_scale}{Boolean (TRUE or FALSE) for whether to scale the raw data with SD = 1}

\item{lower_bound}{the smallest number of profiles in the range of number of profiles to explore; defaults to 2}

\item{upper_bound}{the largest number of profiles in the range of number of profiles to explore; defaults to 9}

\item{r_squared_table}{if TRUE, then a table, rather than a plot, is returned; defaults to FALSE}
}
\value{
A list containing a ggplot2 object and a tibble for the R^2 values
}
\description{
# Create profiles
}
\details{
Create profiles using cluster analysis or mixture modeling
}
\examples{
df <- mtcars
plot_r_squared(df, mpg, wt, hp, qsec, to_center = TRUE, lower_bound = 2, upper_bound = 4)
}
