% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_functions.R
\name{cross_validate}
\alias{cross_validate}
\title{Function to cross-validate the cluster solution using split half or other cross validation}
\usage{
cross_validate(prepared_data, output, variable_names, k)
}
\arguments{
\item{prepared_data}{output from the prepare_data() function}

\item{output}{output from the create_profiles() function}

\item{variable_names}{optional names for clustered variables}

\item{k}{integer for the number of cross-validation attempts}
}
\description{
Function to cross-validate the cluster solution using split half or other cross validation
}
\details{
Function to cross-validate the cluster solution using split half or other cross validation
}

