% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_functions.R
\name{cross_validate}
\alias{cross_validate}
\title{Function to cross-validate the cluster solution using split half or other cross validation}
\usage{
cross_validate(prepared_data, output, variable_vector, cluster_vector, k,
  print_status = T)
}
\arguments{
\item{cluster_assignments}{cluster assignments from calculate_stats() function, in particular the fifth list item from its output}
}
\description{
Function to cross-validate the cluster solution using split half or other cross validation
}
\details{
Function to cross-validate the cluster solution using split half or other cross validation
}

