% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_new.R
\name{cross_validate}
\alias{cross_validate}
\title{Returns statistics from double-split cross validation}
\usage{
cross_validate(df, ..., to_center = FALSE, to_scale = FALSE, n_profiles,
  distance_metric = "squared_euclidean", linkage = "complete", k = 30,
  lower_bound = 2, upper_bound = 9)
}
\arguments{
\item{...}{Additional arguments}

\item{x}{A `prcr` object}
}
\value{
A ggplot2 object
}
\description{
Returns statistics from double-split cross validation
}
\details{
Performs double-split cross validation and returns Cohen's Kappa and percentage agreement statistics.
}
