% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.prcr}
\alias{plot.prcr}
\title{Return plot of cluster centroids}
\usage{
\method{plot}{prcr}(x)
}
\arguments{
\item{x}{A `prcr` object}
}
\value{
A ggplot2 object
}
\description{
Return plot of cluster centroids
}
\details{
Returns ggplot2 plot of cluster centroids
}
