% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_new.R
\name{plot_profiles}
\alias{plot_profiles}
\title{Return plot of cluster centroids}
\usage{
plot_profiles(x, ...)
}
\arguments{
\item{x}{A `prcr` object}

\item{...}{Additional arguments}
}
\value{
A ggplot2 object
}
\description{
Return plot of cluster centroids
}
\details{
Returns ggplot2 plot of cluster centroids
}
