% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_clusters.R
\name{print.prcr}
\alias{print.prcr}
\title{Prints details of prcr cluster solution}
\usage{
\method{print}{prcr}(x, ...)
}
\arguments{
\item{x}{A `prcr` object}

\item{...}{Additional arguments}
}
\description{
Prints details of prcr cluster solution
}
\details{
Prints details of of prcr cluster solution
}
