% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_profiles_cluster.R
\name{summary.prcr}
\alias{summary.prcr}
\title{Concise summary of prcr cluster solution}
\usage{
\method{summary}{prcr}(object, ...)
}
\arguments{
\item{object}{A `prcr` object}

\item{...}{Additional arguments}
}
\description{
Concise summary of prcr cluster solution
}
\details{
Prints a concise summary of prcr cluster solution
}
